Update domain
* Updated domain with missing or incorrect variables
* Updated factory and factory tests that were broken with change

+++ src/main/java/za/ac/cput/factory/TutorProfileFactory.java
@@ -1,5 +1,6 @@
 package za.ac.cput.factory;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import za.ac.cput.domain.TutorProfile;
@@ -8,7 +9,7 @@ import za.ac.cput.util.HelperUtil;
 
 public class TutorProfileFactory {
 
-    public TutorProfile createTutorProfile(User user, String bio, int yearsExperience, double hourlyRate,
+    public TutorProfile createTutorProfile(User user, String bio, int yearsExperience, BigDecimal hourlyRate,
             double averageRating, LocalDateTime createdAt) {
 
         if (user == null || createdAt == null || HelperUtil.isNullOrEmpty(bio)) {
